moved handle method to class CommandHandler in line with the pattern Strategy
