--- client/src/main/java/com/porejemplo/nube/client/StartClient.java
@@ -1,7 +1,0 @@
-package com.porejemplo.nube.client;
-
-public class StartClient {
-    public static void main(String[] args) {
-        new ConsoleClient();
-    }
-}
